pyproject.toml
[project]
name = "maze-agent"
version = "0.1.0"
description = "Autonomous maze-solving agent using Claude Code"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
maze-agent = "maze_agent.cli.main:main"
solve = "maze_agent.cli.main:main"
maze-list = "maze_agent.cli.list_mazes:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py313"
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.bandit]
skips = []
exclude_dirs = ["tests", "docs"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

